requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
joblib
openai
python-dotenv
